Split PMA automation store operations
